Create SensorDataStore
Ethereum smart contract is used to secure store  the condensed data on the blockchain with access control.  Only the authorized user is permitted to access the condensed data. Three levels of authorization for different accounts (read, write, and admin) were used to store and retrieve the data on the blockchain.  A tag-based secure data-store mechanism  is used to minimize the gas consumption of the smart contract design.